Handle default Script properties better
Instead of creating a `defaults` dictionary buried inside of a method and
updating it with embedded properties, we're being a bit smarter.

The Script's ClassName is now set when instansiating. This also opens up
the possibility to create LocalScripts and ModuleScripts progmatically.
Before you would have to set the embedded properties if you wanted either
of those.

And the Name is roughly the same, just instead of using the built-in
self.name of the parent class, we're doing a bit more work. This is for a
later update where Script will be independant of the fs module.