Handle files and folders seperately from compilers
This introduces "processors", which are what we now use to determine what
happens when the compiler encounters a folder or file.

BaseProcessor is very simple, it's all Elixir really needs. But now it's
extremely easy to make other processors, which can be used for more fine
tuning.

For example, something like NevermoreEngine requires a lot of special file
handling. With processers, it's super easy to set everything up.