Make things easier to test
Processors now take the name of the file and its contents, rather than a
path to the file. The is_module function also just takes the file contents
now, we only ever open the path in one place.

This allows us to just use strings when testing, rather than mocking a
file (For the most part. We'll still need to mock a file to test the
compilers, but that's gonna come later).