Move _get_script_class() to the global scope
There's no real reason to have it inside the class, it's not needed outside
the context of this module.

It also doesn't feel like it belongs as a private method. It feels directly
related to process_script(), so I didn't want it by itself at the top of the
class to organize based on privacy. Alone it somehow feels right though.