Spruce up Nevermore's script processing
There's less repetition now. Before we were returning a new ScriptElement,
which all have a `name` and `source` argument.

Since the we only need to disable Scripts that aren't the engine loader,
this frees us up to offload everything to the super, then disable the
script if we need to.